requirements.txt
attrs==20.3.0
autopep8==1.5.7
certifi==2020.12.5
chardet==4.0.0
idna==2.10
iniconfig==1.1.1
kafka-python==2.0.2
packaging==20.9
pluggy==0.13.1
psycopg3 @ git+https://github.com/psycopg/psycopg3.git@45d04a6027888d2a73cb45f118365647a51dbecf#subdirectory=psycopg3
py==1.10.0
pycodestyle==2.7.0
pykka==3.0.0
pyparsing==2.4.7
pytest==6.2.3
pytest-mock==3.6.0
requests==2.25.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
